Oh atlas.. (#846)
* upgrade ent

* Change uints to int, and optional token status

* make ent

* go get google.golang.org/grpc/internal/transport@v1.67.1

* modify baseline

* Add optional field on the thing

* go mod tidy

* post db migration changes

* gofumpt

+++ spark/so/ent/tokenleaf_extension.go
@@ -56,7 +56,7 @@ func FetchInputLeaves(ctx context.Context, leavesToSpend []*pb.TokenLeafToSpend)
 
 		var foundLeaf *TokenLeaf
 		for _, createdLeaf := range receipt.Edges.CreatedLeaf {
-			if createdLeaf.LeafCreatedTransactionOutputVout == leaf.PrevTokenTransactionLeafVout {
+			if createdLeaf.LeafCreatedTransactionOutputVout == int32(leaf.PrevTokenTransactionLeafVout) {
 				foundLeaf = createdLeaf
 				break
 			}
